requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.6
django-utils-six==2.0
djangorestframework==3.12.2
drf-writable-nested==0.6.2
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
static3==0.7.0
gunicorn==20.0.4
psycopg2==2.8.6